look and say pattern
Look and Say Pattern
Difficulty: MediumAccuracy: 62.54%Submissions: 30K+Points: 4
Given an integer n. Return the nth row of the following look-and-say pattern.
1
11
21
1211
111221
Look-and-Say Pattern:

To generate a member of the sequence from the previous member, read off the digits of the previous member, counting the number of digits in groups of the same digit. For example:

1 is read off as "one 1" or 11.
11 is read off as "two 1s" or 21.
21 is read off as "one 2, then one 1" or 1211.
1211 is read off as "one 1, one 2, then two 1s" or 111221.
111221 is read off as "three 1s, two 2s, then one 1" or 312211.
Example 1:

Input:
n = 5
Output: 111221
Explanation: The 5th row of the given pattern
will be 111221.




class Solution
{
  public:
    string lookandsay(int n) {
        // code here
         string str = "1";
        for(int i=1;i<n;i++)
        {
            int count =1;
            string temp = "";
            for(int j=1;j<str.length();j++)
            {
                if(str[j] == str[j-1])
                {
                    count++;
                }
                else{
                    temp += to_string(count)+str[j-1];
                    count =1;
                }
            }
            
            temp+=to_string(count)+str.back();
            str = temp;
        }
        
        return str;
    }
};