Check number present in AP series
Arithmetic Number
Difficulty: EasyAccuracy: 16.63%Submissions: 179K+Points: 2
Given three integers  'A' denoting the first term of an arithmetic sequence , 'C' denoting the common difference of an arithmetic sequence and an integer 'B'. you need to tell whether 'B' exists in the arithmetic sequence or not. Return 1 if B is present in the sequence. Otherwise, returns 0.

Example 1:

Input: A = 1, B = 3, C = 2
Output: 1
Explaination: 3 is the second term of the 
sequence starting with 1 and having a common 
difference 2.


class Solution{
public:
    int inSequence(int A, int B, int C){
        // code here
        
        if(C==0)
          return A==B;
          
        else if((B-A)%C==0 and(B-A)/C>=0){
            return 1;
            
        }
        else{
            return 0;
        }
        
    }
};